Always separate buffer and current data.
It would make sense not to have a separate buffer if buffer_size==1,
but that means memcpy() in write() would need a check (although it
seems to work without it as well). In the interests of more simpler
code, it is better to always copy stuff into a buffer. The idea is
anyway that the class should act as a reasonable IO buffer.