Add a function to parse event parameters.
This is the first step in overhauling the way event parameters are
fed to the simulation. It uses a new format, which is more concise.
This will allow for multiple events be fed on the command line
and could also be used for storing an event list in a file.

Also add the necessary supporting stuff (example program, exceptions
etc.